+++ andes/core/model.py
@@ -1084,7 +1084,7 @@ class Model:
             if instance.has_check_var:
                 instance.check_var(dae_t=dae_t, niter=niter, err=err)
 
-    def l_check_eq(self):
+    def l_check_eq(self, init=False, **kwargs):
         """
         Call the ``check_eq`` method of discrete components to update equation-dependent flags.
 
@@ -1095,9 +1095,17 @@ class Model:
         -------
         None
         """
-        for instance in self.discrete.values():
-            if instance.has_check_eq:
-                instance.check_eq()
+        if init:
+            for instance in self.discrete.values():
+                if instance.has_check_eq:
+                    instance.check_eq(allow_adjust=self.config.allow_adjust,
+                                      adjust_lower=self.config.adjust_lower,
+                                      adjust_upper=self.config.adjust_upper,
+                                      )
+        else:
+            for instance in self.discrete.values():
+                if instance.has_check_eq:
+                    instance.check_eq()
 
     def s_update(self):
         """
@@ -1754,6 +1762,14 @@ class Model:
 
         return f'{self.class_name} ({self.n} {dev_text}) at {hex(id(self))}'
 
+    def _log_init_debug(self, *args):
+        """
+        Helper function to log initialization debug message.
+        """
+
+        if self.system.options.get("init") is True:
+            logger.debug(*args)
+
     def init(self, routine, debug=False):
         """
         Numerical initialization of a model.
@@ -1775,99 +1791,101 @@ class Model:
         else:
             do_init = getattr(self.flags, flag_name)
 
-        logger.debug('%s has <%s> = %s', self.class_name, flag_name, do_init)
+        self._log_init_debug('%s has <%s> = %s', self.class_name, flag_name, do_init)
 
         if do_init:
             kwargs = self.get_inputs(refresh=True)
 
-            logger.debug('%s: initialization sequence:', self.class_name)
-            logger.debug(' -> '.join([str(i) for i in self.calls.init_seq]))
-            logger.debug("%s: assignment initialization phase begins", self.class_name)
+            self._log_init_debug('%s: initialization sequence:', self.class_name)
+            self._log_init_debug(' -> '.join([str(i) for i in self.calls.init_seq]))
+            self._log_init_debug("%s: assignment initialization phase begins", self.class_name)
 
             for idx, name in enumerate(self.calls.init_seq):
                 # single variable - do assignment
                 if isinstance(name, str):
-                    logger.debug("%s: entering <%s> assignment init", self.class_name, name)
+                    self._log_init_debug("%s: entering <%s> assignment init", self.class_name, name)
                     instance = self.__dict__[name]
                     if instance.discrete is not None:
-                        logger.debug("%s: evaluate discrete <%s>", name, instance.discrete)
+                        self._log_init_debug("%s: evaluate discrete <%s>", name, instance.discrete)
                         _eval_discrete(instance, self.config.allow_adjust,
                                        self.config.adjust_lower, self.config.adjust_upper)
 
                     if instance.v_str is not None:
                         arg_print = OrderedDict()
-                        if logger.level <= 10:
+                        if self.system.options.get("init"):
                             for a, b in zip(self.calls.ia_args[name], self.ia_args[name]):
                                 arg_print[a] = b
 
                         if not instance.v_str_add:
                             # assignment is for most variable initialization
-                            logger.debug("%s: new values will be assigned (=)", name)
+                            self._log_init_debug("%s: new values will be assigned (=)", name)
                             instance.v[:] = self.calls.ia[name](*self.ia_args[name])
 
                         else:
                             # in-place add initial values.
                             # Voltage compensators can set part of the `v` of exciters.
                             # Exciters will then set the bus voltage part.
-                            logger.debug("%s: new values will be in-place added (+=)", name)
+                            self._log_init_debug("%s: new values will be in-place added (+=)", name)
                             instance.v[:] += self.calls.ia[name](*self.ia_args[name])
 
                         arg_print[name] = instance.v
 
-                        if logger.level <= 10:
+                        if self.system.options.get("init"):
                             for key, val in arg_print.items():
                                 if isinstance(val, (int, float, np.floating, np.integer)) or \
                                         isinstance(val, np.ndarray) and val.ndim == 0:
 
                                     arg_print[key] = val * np.ones_like(instance.v)
+                                if isinstance(val, np.ndarray) and val.dtype == complex:
+                                    arg_print[key] = [str(i) for i in val]
 
                             tab = Tab(title="v_str of %s is '%s'" % (name, instance.v_str),
                                       header=["idx", *self.calls.ia_args[name], name],
                                       data=list(zip(self.idx.v, *arg_print.values())),
                                       )
-                            logger.debug(tab.draw())
+                            self._log_init_debug(tab.draw())
 
                     # single variable iterative solution
                     if name in self.calls.ii:
-                        logger.debug("\n%s: entering <%s> iterative init", self.class_name,
-                                     pprint.pformat(name))
+                        self._log_init_debug("\n%s: entering <%s> iterative init", self.class_name,
+                                             pprint.pformat(name))
 
                         self.solve_iter(name, kwargs)
 
-                        logger.debug("%s new values are %s", name, self.__dict__[name].v)
+                        self._log_init_debug("%s new values are %s", name, self.__dict__[name].v)
 
                 # multiple variables, iterative
                 else:
-                    logger.debug("\n%s: entering <%s> iterative init",
-                                 self.class_name, name)
+                    self._log_init_debug("\n%s: entering <%s> iterative init",
+                                         self.class_name, name)
 
                     for vv in name:
                         instance = self.__dict__[vv]
 
                         if instance.discrete is not None:
-                            logger.debug("%s: evaluate discrete <%s>",
-                                         name, instance.discrete)
+                            self._log_init_debug("%s: evaluate discrete <%s>",
+                                                 name, instance.discrete)
 
                             _eval_discrete(instance, self.config.allow_adjust,
                                            self.config.adjust_lower, self.config.adjust_upper)
                         if instance.v_str is not None:
-                            logger.debug("%s: v_str = %s", vv, instance.v_str)
+                            self._log_init_debug("%s: v_str = %s", vv, instance.v_str)
 
                             arg_print = OrderedDict()
-                            if logger.level <= 10:
+                            if self.system.options.get("init"):
                                 for a, b in zip(self.calls.ia_args[vv], self.ia_args[vv]):
                                     arg_print[a] = b
-                                logger.debug(pprint.pformat(arg_print))
+                                self._log_init_debug(pprint.pformat(arg_print))
 
                             instance.v[:] = self.calls.ia[vv](*self.ia_args[vv])
 
-                    logger.debug("\n%s: entering <%s> iterative init", self.class_name,
-                                 pprint.pformat(name))
+                    self._log_init_debug("\n%s: entering <%s> iterative init", self.class_name,
+                                         pprint.pformat(name))
                     self.solve_iter(name, kwargs)
 
                     for vv in name:
                         instance = self.__dict__[vv]
-                        logger.debug("%s new values are %s", vv, instance.v)
+                        self._log_init_debug("%s new values are %s", vv, instance.v)
 
             # call custom variable initializer after generated init
             kwargs = self.get_inputs(refresh=True)
@@ -1883,6 +1901,9 @@ class Model:
         Solve iterative initialization.
         """
         for pos in range(self.n):
+            logger.debug("%s: iterative init for %s, device pos = %s",
+                         self.class_name, name, pos)
+
             self.solve_iter_single(name, kwargs, pos)
 
     def solve_iter_single(self, name, inputs, pos):
@@ -1909,26 +1930,39 @@ class Model:
         solved = False
 
         while niter < maxiter:
+            logger.debug("iteration %s:", niter)
+
             i_args = [item[pos] for item in ii_args]  # all variables of one device at a time
             j_args = [item[pos] for item in ij_args]
 
             b = np.ravel(rhs(*i_args))
             A = jac(*j_args)
-            inc = - sp.linalg.lu_solve(sp.linalg.lu_factor(A), b)
-            if np.isnan(inc).any():
-                logger.warning(f"{self.class_name}: nan detected in iterations for {self.idx.v[pos]}.")
-                break
 
-            x0 += inc
-            for idx, item in enumerate(name):
-                inputs[item][pos] = x0[idx]
+            logger.debug("A:\n%s", A)
+            logger.debug("b:\n%s", b)
 
-            mis = np.max(np.abs(inc))
-            niter += 1
-
+            mis = np.max(np.abs(b))
             if mis <= eps:
                 solved = True
                 break
+
+            inc = - sp.linalg.lu_solve(sp.linalg.lu_factor(A), b)
+            if np.isnan(inc).any():
+                if self.u.v[pos] != 0:
+                    logger.debug("%s: nan ignored in iterations for offline device pos = %s",
+                                 self.class_name, pos)
+                else:
+                    logger.error("%s: nan error detected in iterations for device pos = %s",
+                                 self.class_name, pos)
+                break
+
+            x0 += inc
+            logger.debug("solved x0:\n%s", x0)
+
+            for idx, item in enumerate(name):
+                inputs[item][pos] = x0[idx]
+
+            niter += 1
 
         if solved:
             for idx, item in enumerate(name):