+++ andes/models/exciter/exac4.py
@@ -109,7 +109,7 @@ class EXAC4Model(ExcBase):
         self.HLR = HardLimiter(u=self.LR_y, lower=self.vfmin, upper=self.vfmax,
                                info='Hard limiter on regulator output')
 
-        self.vout.e_str = 'LR_y*HLR_zi + vfmin*HLR_zl + vfmax*HLR_zu - vout'
+        self.vout.e_str = 'ue * (LR_y*HLR_zi + vfmin*HLR_zl + vfmax*HLR_zu) - vout'
 
 
 class EXAC4(EXAC4Data, EXAC4Model):