+++ andes/models/exciter/exst1.py
@@ -125,13 +125,15 @@ class EXST1Model(ExcBase):
         self.HLR = HardLimiter(u=self.WF_y, lower=self.vfmin, upper=self.vfmax,
                                info='Hard limiter on regulator output')
 
-        self.vout.e_str = 'LR_y*HLR_zi + vfmin*HLR_zl + vfmax*HLR_zu - vout'
+        self.vout.e_str = 'ue * (LR_y*HLR_zi + vfmin*HLR_zl + vfmax*HLR_zu) - vout'
 
 
 class EXST1(EXST1Data, EXST1Model):
     """
     EXST1-type static excitation system.
     """
+
+    # TODO: disable offline EXST1 in eigenvalue analysis
 
     def __init__(self, system, config):
         EXST1Data.__init__(self)