+++ tests/test_known_good.py
@@ -1,4 +1,5 @@
 import os
+import logging
 import unittest
 
 import dill
@@ -6,6 +7,8 @@ import numpy as np
 
 import andes
 from andes.utils import get_case
+
+logger = logging.getLogger(__name__)
 
 
 class TestKnownResults(unittest.TestCase):
@@ -18,7 +21,7 @@ class TestKnownResults(unittest.TestCase):
             (('ieee14/ieee14.raw', 'ieee14/ieee14.dyr'), 'ieee14_2s.pkl', 2),
             )
 
-    def tnc(self, case_path, pkl_path, tf):
+    def tnc(self, case_path, pkl_path, tf, pkl_prefix='pkl'):
         """
         Test and compare
         """
@@ -29,7 +32,10 @@ class TestKnownResults(unittest.TestCase):
         else:
             case_path = get_case(case_path)
 
-        ss = compare_results(case_path, pkl_path, tf=tf, addfile=addfile)
+        ss = compare_results(case_path,
+                             os.path.join(pkl_prefix, pkl_path),
+                             tf=tf,
+                             addfile=addfile)
 
         self.assertEqual(ss.exit_code, 0, "Exit code is not 0.")
 
@@ -48,10 +54,9 @@ def compare_results(case, pkl_name, addfile=None, tf=10):
     ss.PFlow.run()
 
     ss.TDS.config.tstep = 1/30
-    ss.TDS.config.tol = 1e-6
+    ss.TDS.config.tol = 1e-4
     ss.TDS.config.fixt = 1
     ss.TDS.config.shrinkt = 0
-    ss.TDS.config.honest = 0
     ss.TDS.run()
 
     test_dir = os.path.dirname(__file__)
@@ -62,9 +67,15 @@ def compare_results(case, pkl_name, addfile=None, tf=10):
     indices = np.hstack((ss.GENROU.omega.a,
                          ss.dae.n + ss.GENROU.tm.a,
                          ss.dae.n + ss.GENROU.vf.a))
+    logger.info(case)
+    logger.info("This results: %s", ss.dae.xy[indices])
+    logger.info("Known results: %s", results)
 
     np.testing.assert_almost_equal(ss.dae.xy[indices],
                                    results,
                                    err_msg=f"{case} test error",
-                                   decimal=3)
+                                   decimal=2,)
+    # actual algeb. errors will be `tstep` times greater.
+    # combined with a tol of 1e-4, accuracy is to the 2nd decimal place.
+
     return ss